% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year-progress.R
\name{year_progress}
\alias{year_progress}
\title{Give the progress of the year for a given date}
\usage{
year_progress(date, is_leap_year = FALSE)
}
\arguments{
\item{date}{A date}

\item{is_leap_year}{Is that year a leap year?}
}
\description{
Rajouter une petite description
}
\examples{
year_progress(Sys.Date())
}
